# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample CRUD api for estuary db",
        "title": "Sample CRUD api for estuary db",
        "contact": {
            "name": "Me",
            "url": "http://me.com/terms.html",
            "email": "alvin@protocol.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/contentdealproposalparameterslogs": {
            "get": {
                "description": "GetAllContentDealProposalParametersLogs is a handler to get a slice of record(s) from content_deal_proposal_parameters_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalParametersLogs"
                ],
                "summary": "Get list of ContentDealProposalParametersLogs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page requested (defaults to 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of records in a page  (defaults to 20)",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "db sort order column",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PagedResults"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "add to add a single record to content_deal_proposal_parameters_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalParametersLogs"
                ],
                "summary": "Add an record to content_deal_proposal_parameters_logs table",
                "parameters": [
                    {
                        "description": "Add ContentDealProposalParametersLogs",
                        "name": "ContentDealProposalParametersLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/contentdealproposalparameterslogs/{contentDealProposalParametersLogsID}": {
            "get": {
                "description": "GetContentDealProposalParametersLogs is a function to get a single record from the content_deal_proposal_parameters_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalParametersLogs"
                ],
                "summary": "Get record from table ContentDealProposalParametersLogs by  contentDealProposalParametersLogsID",
                "operationId": "contentDealProposalParametersLogsID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalParametersLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "ErrNotFound, db record for id not found - returns NotFound HTTP 404 not found error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single record from content_deal_proposal_parameters_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalParametersLogs"
                ],
                "summary": "Update an record in table content_deal_proposal_parameters_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalParametersLogsID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ContentDealProposalParametersLogs record",
                        "name": "ContentDealProposalParametersLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single record from content_deal_proposal_parameters_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalParametersLogs"
                ],
                "summary": "Delete a record from content_deal_proposal_parameters_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalParametersLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalParametersLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "api.PagedResults": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
...
        "model.ContentDealProposalParametersLogs": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "[ 1] content                                        INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "createdAt": {
                    "description": "[13] created_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                },
                "deltaNodeUUID": {
                    "description": "[15] delta_node_uuid                                TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "duration": {
                    "description": "[ 3] duration                                       INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "endEpoch": {
                    "description": "[ 5] end_epoch                                      INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "id": {
                    "description": "[ 0] id                                             INT8                 null: false  primary: true   isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "type": "integer"
                },
                "label": {
                    "description": "[ 2] label                                          TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "nodeInfo": {
                    "description": "[ 9] node_info                                      TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "removeUnsealedCopy": {
                    "description": "[ 7] remove_unsealed_copy                           BOOL                 null: true   primary: false  isArray: false  auto: false  col: BOOL            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "requesterInfo": {
                    "description": "[10] requester_info                                 TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "requestingAPIKey": {
                    "description": "[11] requesting_api_key                             TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "skipIPNiAnnounce": {
                    "description": "[ 8] skip_ip_ni_announce                            BOOL                 null: true   primary: false  isArray: false  auto: false  col: BOOL            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "startEpoch": {
                    "description": "[ 4] start_epoch                                    INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "systemContentDealProposalParametersID": {
                    "description": "[12] system_content_deal_proposal_parameters_id     INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "transferParams": {
                    "description": "[ 6] transfer_params                                TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "updatedAt": {
                    "description": "[14] updated_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}